+++ src/languages/QOwnNotes_da.ts
@@ -2965,14 +2965,29 @@ hvis papirkurven er aktiveret på din ownCloud server du skal kunne gendanne dem
 <context>
     <name>NotePreviewWidget</name>
     <message>
-        <location filename="../widgets/notepreviewwidget.cpp" line="236"/>
+        <location filename="../widgets/notepreviewwidget.cpp" line="237"/>
         <source>Copy image file path</source>
         <translation>Kopier billedfilsti</translation>
     </message>
     <message>
-        <location filename="../widgets/notepreviewwidget.cpp" line="240"/>
+        <location filename="../widgets/notepreviewwidget.cpp" line="241"/>
         <source>Copy link location</source>
         <translation>Kopiér linkplacering</translation>
+    </message>
+    <message>
+        <location filename="../widgets/notepreviewwidget.cpp" line="244"/>
+        <source>Export generated raw HTML</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../widgets/notepreviewwidget.cpp" line="277"/>
+        <source>HTML files</source>
+        <translation type="unfinished">HTML-filer</translation>
+    </message>
+    <message>
+        <location filename="../widgets/notepreviewwidget.cpp" line="278"/>
+        <source>Export preview as raw HTML file</source>
+        <translation type="unfinished"></translation>
     </message>
 </context>
 <context>