+++ src/languages/QOwnNotes_el.ts
@@ -2965,14 +2965,29 @@ If the trash is enabled on your ownCloud server you should be able to restore th
 <context>
     <name>NotePreviewWidget</name>
     <message>
-        <location filename="../widgets/notepreviewwidget.cpp" line="236"/>
+        <location filename="../widgets/notepreviewwidget.cpp" line="237"/>
         <source>Copy image file path</source>
         <translation>Αντιγραφή διαδρομής αρχείου εικόνας</translation>
     </message>
     <message>
-        <location filename="../widgets/notepreviewwidget.cpp" line="240"/>
+        <location filename="../widgets/notepreviewwidget.cpp" line="241"/>
         <source>Copy link location</source>
         <translation>Αντιγράψτε τη θέση σύνδεσης</translation>
+    </message>
+    <message>
+        <location filename="../widgets/notepreviewwidget.cpp" line="244"/>
+        <source>Export generated raw HTML</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../widgets/notepreviewwidget.cpp" line="277"/>
+        <source>HTML files</source>
+        <translation type="unfinished">Αρχεία HTML</translation>
+    </message>
+    <message>
+        <location filename="../widgets/notepreviewwidget.cpp" line="278"/>
+        <source>Export preview as raw HTML file</source>
+        <translation type="unfinished"></translation>
     </message>
 </context>
 <context>